+++ app/src/main/AndroidManifest.xml
@@ -11,7 +11,7 @@
         android:roundIcon="@mipmap/ic_launcher_round"
         android:supportsRtl="true"
         android:theme="@style/AppTheme">
-        <activity android:name=".CreateWorkoutActivity"></activity>
+        <activity android:name=".ui.createWorkout.CreateWorkoutActivity"></activity>
         <activity
             android:name=".WorkoutsActivity"
             android:label="@string/title_activity_workouts"
