+++ app/src/main/java/hu/bme/aut/fitary/ui/barChart/BarChartFragment.kt
@@ -1,4 +1,4 @@
-package hu.bme.aut.fitary.ui.charts
+package hu.bme.aut.fitary.ui.barChart
 
 import android.os.Bundle
 import android.view.LayoutInflater
