+++ app/src/main/java/hu/bme/aut/fitary/ui/socialWorkouts/SocialWorkoutsFragment.kt
@@ -1,4 +1,4 @@
-package hu.bme.aut.fitary.ui.workouts
+package hu.bme.aut.fitary.ui.socialWorkouts
 
 import android.os.Bundle
 import android.view.LayoutInflater
