+++ app/src/main/res/layout/fragment_chart_bar.xml
@@ -4,7 +4,7 @@
     xmlns:tools="http://schemas.android.com/tools"
     android:layout_width="match_parent"
     android:layout_height="match_parent"
-    tools:context=".ui.charts.BarChartFragment">
+    tools:context=".ui.barChart.BarChartFragment">
 
     <com.github.mikephil.charting.charts.HorizontalBarChart
         android:id="@+id/barChart"
