+++ app/src/main/res/layout/fragment_chart_pie.xml
@@ -5,7 +5,7 @@
     xmlns:tools="http://schemas.android.com/tools"
     android:layout_width="match_parent"
     android:layout_height="match_parent"
-    tools:context=".ui.charts.PieChartFragment">
+    tools:context=".ui.pieChart.PieChartFragment">
 
     <com.github.mikephil.charting.charts.PieChart
         android:id="@+id/pieChart"
