+++ app/src/main/res/layout/fragment_workouts_social.xml
@@ -4,7 +4,7 @@
     xmlns:tools="http://schemas.android.com/tools"
     android:layout_width="match_parent"
     android:layout_height="match_parent"
-    tools:context=".ui.workouts.SocialWorkoutsFragment">
+    tools:context=".ui.socialWorkouts.SocialWorkoutsFragment">
 
     <androidx.recyclerview.widget.RecyclerView
         android:id="@+id/rvSocialWorkouts"
