+++ app/src/main/res/layout/fragment_workouts_user.xml
@@ -5,7 +5,7 @@
     xmlns:tools="http://schemas.android.com/tools"
     android:layout_width="match_parent"
     android:layout_height="match_parent"
-    tools:context=".ui.workouts.UserWorkoutsFragment">
+    tools:context=".ui.userWorkouts.UserWorkoutsFragment">
 
     <androidx.recyclerview.widget.RecyclerView
         android:id="@+id/rvUserWorkouts"
