+++ app/src/main/res/navigation/mobile_navigation.xml
@@ -7,13 +7,13 @@
 
     <fragment
         android:id="@+id/nav_user_workouts"
-        android:name="hu.bme.aut.fitary.ui.workouts.UserWorkoutsFragment"
+        android:name="hu.bme.aut.fitary.ui.userWorkouts.UserWorkoutsFragment"
         android:label="@string/menu_my_workouts"
         tools:layout="@layout/fragment_workouts_user" />
 
     <fragment
         android:id="@+id/nav_social_workouts"
-        android:name="hu.bme.aut.fitary.ui.workouts.SocialWorkoutsFragment"
+        android:name="hu.bme.aut.fitary.ui.socialWorkouts.SocialWorkoutsFragment"
         android:label="@string/menu_social_workouts"
         tools:layout="@layout/fragment_workouts_social" />
